mvp settings
