+++ app/src/main/java/com/example/android/playlistmaker/settings/presentation/SettingsView.kt
@@ -1,0 +1,4 @@
+package com.example.android.playlistmaker.settings.presentation
+
+interface SettingsView {
+}
